requirements.txt
torch
torchvision
torchaudio
opencv-python
ultralytics
timm
scikit-learn
pyyaml
fastapi
uvicorn
gradio
numpy
pandas
python-multipart
onnx
onnxruntime
python-dotenv
